REF: move all viewer code to the viewer submodule
