+++ bluesky_browser/viewer/hints.py
